requirements.txt
torch
torchvision
numpy
pandas
opencv-python
matplotlib
Pillow

ultralytics

transformers
timm

#helium
bs4